requirements.txt
python-dotenv~=1.0.0
cryptography~=3.4.8
flask~=3.0.0
supabase~=1.1.1
requests~=2.27.1
bs4~=0.0.1
beautifulsoup4~=4.11.1